requirements.txt
pyrogram==2.0.106
python-dotenv==1.0.1
prettytable==3.12.0
TgCrypto==1.2.5
aiohttp==3.11.10